fixed html validation problems
Added xml declaration to the html documents.

Added the tbody tag for each table, which is required for strict xhtml
validation.

Fixed a problem with one of the tables where a tr-endtag were missing,
basically.

Also center aligned the whole design, in some browsers any way.

Also fixed a charset problem, where the charset was declared as utf8
instead of utf-8.